requirements.txt
numpy
opencv-python
Pillow
safetensors
torch
transformers
onnx
tensorrt
